build: Add croot function to build environment script
This commit introduces a `croot` function to the build environment script, providing a convenient way to navigate to the root of the build tree.

Similar to the `croot` command found in the Android build environment, this function allows developers to quickly return to the top level of the source tree from any location within it.

This addition improves the usability and familiarity of the script, especially for those accustomed to working with Android's build system.