feat(tests): Introduce compatibility_suites to allow global suites in bp
Introduce a new feature that allows compatibility suites to be specified globally in module_info.bp.

This change adds a new 'compatibility_suites' field to module_info.bp, which can be used to specify a list of compatibility suites that should be included in the build.
These suites will be added to the global list of compatibility suites, which is used to determine which tests should be skipped in the presubmit check.

This feature allows developers to easily enable or disable compatibility suites for a module, without having to modify the presubmit check configuration.