build: definitions: Add function to print variable names and values
This commit adds a new function, `print_vars`, to the `definitions` module. This function takes two arguments: a list of variable names and a dictionary of variable names to values. It then prints the names of the variables and their corresponding values.

This function is useful for debugging purposes, as it allows developers to easily see the values of variables that are used in the build system.