build: definitions: add function to retrieve `module_info.bp` files from subdirectories
Adds a new function, `all_subdir_module_info`, to the `definitions` module.
This function allows users to retrieve a list of `module_info.bp` files in subdirectories of the specified module file directory.

BREAKING CHANGE: The new function has been added to the `definitions` module.
If you are using this module, please update your code to use the new function.