build: definitions: add function to find all directories with a specified name under a list of base directories
Adds a new function to the definitions module called `all_named_dirs_under`. This function takes two arguments: a directory name and a list of base directories. It returns a list of all directories with the specified name under the given base directories.

This function is meant to be used like:

```
INC_DIRS := $(call all-named-dirs-under,inc,.)
```

This will create a variable called `INC_DIRS` that contains a list of all directories with the name `inc` under the current directory.