build: definitions: Add all_subdir_named_dirs() function
Adds a new function to `core/definitions.py` called `all_subdir_named_dirs()`. This function finds all directories under the current directory that have a specified name. This can be useful for tasks such as locating all directories that contain a particular type of file or that are used for a specific purpose.