build: definitions: add all_subdir_named_files() to find all files with a specific name under the current directory
This commit adds a new function, `all_subdir_named_files()`, to the `core` module. This function takes a file name pattern as input and returns a list of all files that match the specified pattern under the current directory.

This function is useful for finding all of the files in a project that have a specific name, such as all of the header files or all of the source files. It can also be used to find all of the files in a directory that have a specific extension, such as all of the Python files or all of the C++ files.