build: definitions: ability to get all cpp and c files under a specified directory
Add a functions to find all of the c and cpp files in the current directory or a specified directory. This is meant to be used to help build a source list when building a recipe.

BREAKING CHANGE: The LOCAL_CPP_EXTENSION variable is now respected when searching for all cpp files. This means that if you have a cpp file with a different extension, it will not be found by the all_cpp_files() function.