refactor(core): Add find_S_files API
This commit adds four new file-finding functions:
- all_S_files_under()
- all_html_files_under()
- all_subdir_html_files()
- find_subdir_files()

These functions allow users to find all files of a given type under a specified directory or list of directories. The functions are designed to be flexible and easy to use, and they can be used to find a variety of file types, including source code, test files, and assets.

These functions are useful for automating tasks such as building, testing, and packaging applications. They can also be used to create custom tools and scripts that work with files in a specific directory or set of directories.