build: definitions: find html files in directories
This commit adds a new function, `find_other_html_files`, that finds all `.html` files under the named directories relative to the current path.

This function is useful for finding all of the HTML files in a project, or for finding all of the HTML files that are related to a specific topic.

The function takes a list of directories as input, and returns a list of all of the `.html` files that are found under those directories.

The function uses the `rglob` function to search for `.html` files in the specified directories and their subdirectories. The `rglob` function is a recursive version of the `glob` function, which means that it will search for files in all of the subdirectories of the specified directories.

The `find_other_html_files` function is a useful addition to the `core` module, and it can be used to find all of the HTML files in a project or for finding all of the HTML files that are related to a specific topic.