build: definitions: add function to find files in subdirectories
Adds a new function, `find_files_in_subdirs`, to the `core` package to find files in the given subdirectories relative to the base directory.

This function is useful for finding files that are not in the same directory as the script that is calling the function. For example, it can be used to find all of the `.cpp` files in a project's source tree.