feat: add 'copied_target_license_metadata_rule' to check for missing 'meta_lic'
Adds a new rule, 'copied_target_license_metadata_rule', to check if a target's `meta_lic` attribute is defined.

If the attribute is not defined, it is automatically added to the target's metadata.

This rule is necessary to ensure that all targets have a valid `meta_lic` attribute, which is used to generate license metadata files.