build: definitions: add more non-module-specific functions to definitions
This commit moves several non-module-specific functions from the top level of the definitions module to more specific locations.

This change improves the organization of the definitions module by grouping related functions into logical groups. It also reduces the amount of code in the top-level of the module, which makes it easier to maintain.

The following functions are moved:

- `get_host_2nd_arch` is moved to the `helper` module.
- `touch` is moved to the `utils` module.
- `product_copy_files` is moved to the `utils` module.
- `parse_and_copy_files` is moved to the `utils` module.
- `include` is moved to the `utils` module.