style: add ninja-style color output to test output
This commit adds ninja-style color output to the test output. This makes the test output easier to read and understand.

Additionally, this commit fixes an issue where the test output was not being printed in the correct order. This issue has been fixed by changing the way that the test output is printed.

BREAKING CHANGE: The test output format has changed. This change should not affect the functionality of the tests, but it may require updates to any scripts or tools that rely on the test output.