build: compliance: rework conditionset representation
This change reworks the representation of `ConditionSet` to be a bit more flexible.

Previously, `ConditionSet` was represented as a bitmask, which was not very extensible. This change represents `ConditionSet` as a set of bits, which allows us to easily add new conditions in the future.

Additionally, this change adds a number of new methods to `ConditionSet` that make it easier to work with sets of conditions. These methods include:

- `MatchingAny()`: Returns the subset of `self` equal to any of the given conditions.
- `MatchingAnySet()`: Returns the subset of `self` that are members of any of the given condition sets.
- `HasAny()`: Returns True when `self` contains at least one of the given conditions.
- `MatchesEverySet()`: Returns True when `self` has a non-empty intersection with each of the given condition sets.
- `HasAll()`: Returns True when `self` contains every one of the given conditions.
- `Intersection()`: Returns the subset of `self` that are members of every other set.
- `Minus()`: Returns the subset of `self` that are not equal to any conditions.
- `Difference()`: Returns the subset of `self` that are not members of any other set.
- `Len()`: Returns the number of license conditions in the set.
- `AsList()`: Returns a list of the license conditions in the set.
- `Names()`: Returns a list of the names of the license conditions in the set.
- `IsEmpty()`: Returns True when the set contains no license conditions.

These methods make it much easier to work with sets of conditions, and they will be useful in a variety of future changes.