feat: provision for function granularity in annotating edges
Adds new classes `AnnotationSet` and `TargetEdgeAnnotations` to represent sets of annotations associated with an edge in the license graph.

This change allows us to represent and propagate edge annotations more efficiently and consistently. The annotations can be used to capture various metadata about the edge, such as whether it represents a dynamic link or a derivation.

The following helper functions are also added to assist in edge annotation:

- `edgeIsDynamicLink`
- `edgeIsDerivation`
- `ImpliesRestricted`
- `NewLicenseConditionSet`
- `RecognizedConditionNames`